backend/requirements.txt
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.1
daphne== 2.5.0
# uvicorn
pyOpenSSL
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
service-identity==24.1.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.0.3
django-sslserver==0.22
channels_redis==4.2.0
Twisted[tls,http2]